+++ packages/ricos-styles/web/src/decorations/color-decoration.ts
@@ -1,11 +1,11 @@
 import type { Decoration } from 'ricos-schema';
-import { FontSizeData_fontType, Decoration_Type } from 'ricos-schema';
+import { Decoration_Type } from 'ricos-schema';
 import type { CustomTextualStyle } from 'ricos-types';
 import { EmptyDecoration } from './empty-decoration';
-import type { TextDecoration } from '../../models/decoration';
+import type { TextDecoration } from '../models/decoration';
 
-export class FontSizeDecoration implements TextDecoration {
-  type = Decoration_Type.FONT_SIZE;
+export class ColorDecoration implements TextDecoration {
+  type = Decoration_Type.COLOR;
 
   private readonly customStyle: CustomTextualStyle;
 
@@ -14,31 +14,30 @@ export class FontSizeDecoration implements TextDecoration {
   }
 
   static of(decoration: Decoration) {
-    if (decoration.type !== Decoration_Type.FONT_SIZE) {
+    if (decoration.type !== Decoration_Type.COLOR) {
       throw new TypeError(`invalid decoration initializer ${decoration}`);
     }
 
-    return new FontSizeDecoration(
-      decoration.fontSizeData?.value ? { fontSize: decoration.fontSizeData?.value } : {}
-    );
+    const { foreground: color, background: backgroundColor } = decoration.colorData || {};
+    return new ColorDecoration({ color, backgroundColor });
   }
 
   getDecoration(): Decoration {
     return {
       type: this.type,
-      fontSizeData: {
-        value: parseInt(this.customStyle.fontSize as string),
-        unit: FontSizeData_fontType.PX,
+      colorData: {
+        ...(this.customStyle.color && { foreground: this.customStyle.color }),
+        ...(this.customStyle.backgroundColor && { background: this.customStyle.backgroundColor }),
       },
     };
   }
 
   static fromCustomStyle(customStyle: CustomTextualStyle): TextDecoration {
-    return new FontSizeDecoration(customStyle);
+    return new ColorDecoration(customStyle);
   }
 
   fromCustomStyle(customStyle: CustomTextualStyle): TextDecoration {
-    return FontSizeDecoration.fromCustomStyle(customStyle);
+    return ColorDecoration.fromCustomStyle(customStyle);
   }
 
   toCustomStyle(): CustomTextualStyle {
@@ -46,10 +45,10 @@ export class FontSizeDecoration implements TextDecoration {
   }
 
   overrideWith(decoration: TextDecoration): TextDecoration {
-    if (!(decoration instanceof FontSizeDecoration || decoration instanceof EmptyDecoration)) {
+    if (!(decoration instanceof ColorDecoration || decoration instanceof EmptyDecoration)) {
       throw new TypeError(`invalid merge decoration ${decoration}`);
     }
     const customStyle = { ...this.customStyle, ...decoration.toCustomStyle() };
-    return new FontSizeDecoration(customStyle);
+    return new ColorDecoration(customStyle);
   }
 }