+++ packages/ricos-styles/web/src/document-style/document-style.spec.ts
@@ -1,5 +1,6 @@
 import DocumentStyle from './document-style';
-import { decorations, nodeStyle, textStyle } from '../tests/consts';
+import { decorations, nodeStyle, textStyle, headingNode } from '../tests/test-cases';
+import type { HeadingNode } from 'ricos-content';
 
 describe('Document Style', () => {
   const documentStyle = {
@@ -37,5 +38,31 @@ describe('Document Style', () => {
     expect(
       JSON.stringify(styleTag.props.children).includes('--ricos-custom-p-background-color: #808080')
     ).toBeTruthy();
+  });
+
+  it('Should fromNode match expected', () => {
+    const documentStyleContent = DocumentStyle.fromNode(headingNode as HeadingNode).toContent();
+    const expected = {
+      headerThree: {
+        decorations: [
+          {
+            type: 'COLOR',
+            colorData: {
+              foreground: '#818181',
+            },
+          },
+          {
+            type: 'BOLD',
+            fontWeightValue: 700,
+          },
+        ],
+        nodeStyle: {
+          paddingTop: '5px',
+          paddingBottom: '2px',
+        },
+        lineHeight: '3',
+      },
+    };
+    expect(documentStyleContent).toStrictEqual(expected);
   });
 });