+++ packages/ricos-styles/web/src/document-style/document-style.ts
@@ -3,19 +3,25 @@ import type {
   Decoration_Type,
   DocumentStyle as RichContentDocumentStyle,
 } from 'ricos-schema';
-import { Decorations } from './decorations';
+import { Decorations } from '../decorations';
 import type { TextDecoration } from '../models/decoration';
 import type { DocumentStyle, TextNodeType } from '../models/styles';
 import { TextStyleTransformer } from '../text-style-transformer';
 import RicosTextualTheme from '../textual-theme/textual-theme';
 import TextStyle from './text-style';
 import NodeStyle from './node-style';
+import { TextNodeTransformer } from '../text-node-transformer';
+import type { ParagraphNode, HeadingNode } from 'ricos-content';
 
 export default class RicosDocumentStyle implements DocumentStyle {
   documentStyle: RichContentDocumentStyle;
 
   constructor(documentStyle: RichContentDocumentStyle) {
     this.documentStyle = documentStyle;
+  }
+
+  static fromNode(node: ParagraphNode | HeadingNode): RicosDocumentStyle {
+    return new RicosDocumentStyle(new TextNodeTransformer(node).toDocumentStyle());
   }
 
   getDecoration(nodeType: TextNodeType, decorationType: Decoration_Type): TextDecoration {