+++ packages/ricos-styles/web/src/index.ts
@@ -1,1 +1,2 @@
 export { StylesContext } from './context';
+export { default as RicosDocumentStyle } from './document-style/document-style';